requirements.txt
joblib==0.14.1
llvmlite==0.31.0
lxml==4.5.0
numba==0.48.0
numpy==1.18.1
opencv-contrib-python==4.2.0.32
pandas==0.25.3
Pillow==7.0.0
python-docx==0.8.10
